Fix interaction between assertions and server thread
We can't throw assertion errors without affecting the server's stability, which means we need to move the expectations outside the stubbing of the fixtures.

This commit does all changes necessary and adapts all tests by first setting the server's behavior and then providing means to ask about what's happened.

Everything related to the expected request that the API should have done is made through a new object of type RecordedRequest that the server provides. Everything related to the response must be tested through the actual output of the client call, which supports the idea of these tests being end-to-end.